useLogとuseRollを分離
